# issuer-node/api/api.yaml
openapi: 3.0.0
info:
  title: Privado ID - Issuer
  description: |
    The Issuer Node Core API is ideal for users who need multiple identities and for integrator profiles, who want to 
    create solutions based on Privado ID functionalities and might be interested in having access to low level 
    information such as Merkle Trees.
    The Issuer Node API provide the following functionalities:
    * Create and retrieve Identities
    * Create a Verifiable Credential (VC)
    * Retrieve a Verifiable Credential or a list of Verifiable Credentials
    * Generate JSON to create a QR Code and use that to accept credentials in a wallet
    * Revoke a Verifiable Credential
    * Check revocation status of a Verifiable Credential
    * Retrieve the Revocation Status of a Verifiable Credential
    * Call Agent Endpoint using the Wallet App
    * Handle connections.
  version: "1"
...
paths:
...
  /v2/{identifier}/authentication:
    post:
      summary: Get Authentication Message
      operationId: Authentication
      description: This endpoint returns an authentication message
      tags:
        - Auth
      parameters:
        - $ref: '#/components/parameters/pathIdentifier'
        - name: type
          in: query
          required: false
          description: >
            Type:
              * `link` - (default value) Return a link redirection to the raw content. Easier to scan.   
              * `raw` - Return the raw QR code.
          schema:
            type: string
            enum: [ raw, link ]

      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
...
  /v2/agent:
    post:
      summary: Agent
      operationId: Agent
      description: Identity Agent Endpoint
      tags:
        - Agent
      requestBody:
        required: true
        content:
          text/plain:
            schema:
              type: string
              example: jwz-token
      responses:
        '200':
          description: TBD
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /v1/agent:
    post:
      summary: Agent V1
      operationId: AgentV1
      description: Identity Agent Endpoint for v1.
      deprecated: true
      tags:
        - Agent
      requestBody:
        required: true
        content:
          text/plain:
            schema:
              type: string
              example: jwz-token
      responses:
        '200':
          description: TBD
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
...
    UUIDString:
      type: string
      x-omitempty: false
      example: b7144f1c-d54e-4f67-a4f1-f2e7ff1beb07
...
    GenericErrorMessage:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: 'Something happen'

    #identity
...
    AuthenticationResponse:
      type: object
      required:
        - message
        - sessionID
      properties:
        message:
          type: string
          example: iden3comm://?request_uri=https%3A%2F%2Fissuer-demo.privado.id%2Fapi%2Fqr-store%3Fid%3Df780a169-8959-4380-9461-f7200e2ed3f4
        sessionID:
          $ref: '#/components/schemas/UUIDString'
...
    BasicMessage:
      type: object
      required:
        - body
        - id
        - typ
        - type
        - threadID
        - from
        - to
      properties:
        id:
          type: string
        typ:
          type: string
        type:
          type: string
        threadID:
          type: string
        body:
          type: null
        from:
          type: string
        to:
          type: string
    AgentResponse:
      $ref: '#/components/schemas/BasicMessage'
...
  parameters:
...
    pathIdentifier:
      name: identifier
      in: path
      required: true
      description: Issuer identifier
      schema:
        type: string
...
  responses:
    '400':
      description: 'Bad Request'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericErrorMessage'
...
    '500':
      description: 'Internal Server  error'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericErrorMessage'
...